requirements.txt
alembic==1.7.6
astroid==2.9.3
autopep8==1.6.0
bcrypt==3.2.0
Bootstrap-Flask==2.0.2
cffi==1.15.0
click==8.0.4
dominate==2.6.0
flake8==4.0.1
Flask==2.0.3
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
isort==5.10.1
itsdangerous==2.1.0
Jinja2==3.0.3
lazy-object-proxy==1.7.1
Mako==1.1.6
MarkupSafe==2.1.0
mccabe==0.6.1
platformdirs==2.5.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
pylint==2.12.2
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
pylint-plugin-utils==0.7
python-dotenv==0.19.2
six==1.16.0
SQLAlchemy==1.4.31
sqlalchemy-views==0.3.1
toml==0.10.2
typing_extensions==4.1.1
visitor==0.1.3
Werkzeug==2.0.3
wrapt==1.13.3
WTForms==3.0.1
alembic==1.7.6
astroid==2.9.3
autopep8==1.6.0
bcrypt==3.2.0
Bootstrap-Flask==2.0.2
cffi==1.15.0
click==8.0.4
dominate==2.6.0
flake8==4.0.1
Flask==2.0.3
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
gunicorn==20.1.0
isort==5.10.1
itsdangerous==2.1.0
Jinja2==3.0.3
lazy-object-proxy==1.7.1
Mako==1.1.6
MarkupSafe==2.1.0
mccabe==0.6.1
platformdirs==2.5.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
pylint==2.12.2
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
pylint-plugin-utils==0.7
python-dotenv==0.19.2
six==1.16.0
SQLAlchemy==1.4.31
sqlalchemy-views==0.3.1
toml==0.10.2
typing_extensions==4.1.1
visitor==0.1.3
Werkzeug==2.0.3
wrapt==1.13.3
WTForms==3.0.1
alembic==1.7.6
astroid==2.9.3
autopep8==1.6.0
bcrypt==3.2.0
Bootstrap-Flask==2.0.2
cffi==1.15.0
click==8.0.4
dominate==2.6.0
flake8==4.0.1
Flask==2.0.3
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
gunicorn==20.1.0
isort==5.10.1
itsdangerous==2.1.0
Jinja2==3.0.3
lazy-object-proxy==1.7.1
Mako==1.1.6
MarkupSafe==2.1.0
mccabe==0.6.1
platformdirs==2.5.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
pylint==2.12.2
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
pylint-plugin-utils==0.7
python-dotenv==0.19.2
six==1.16.0
SQLAlchemy==1.4.31
sqlalchemy-views==0.3.1
toml==0.10.2
typing_extensions==4.1.1
visitor==0.1.3
Werkzeug==2.0.3
wrapt==1.13.3
WTForms==3.0.1
